Finishing The Hearing Sense State
Now the AI can Hear the Sound of the Jump of the Player within certain radius and Will Go Investigate.
If the AI Saw the Player Then it will change to Chasing State , And if the Player went out the sight radius then It AI will Go to Patroling State.